+++ tests/modeling/test_results.py
@@ -49,7 +49,13 @@ def test_calculate_individual_shrinkage(load_model_for_test, testdata):
 def test_calculate_individual_parameter_statistics(load_model_for_test, testdata):
     model = load_model_for_test(testdata / 'nonmem' / 'secondary_parameters' / 'pheno.mod')
     rng = np.random.default_rng(103)
-    stats = calculate_individual_parameter_statistics(model, 'CL/V', rng=rng)
+    stats = calculate_individual_parameter_statistics(
+        model,
+        'CL/V',
+        model.modelfit_results.parameter_estimates,
+        model.modelfit_results.covariance_matrix,
+        rng=rng,
+    )
 
     assert stats['mean'][0] == pytest.approx(0.004700589484324183)
     assert stats['variance'][0] == pytest.approx(8.086653508585209e-06)
@@ -57,14 +63,26 @@ def test_calculate_individual_parameter_statistics(load_model_for_test, testdata
 
     model = load_model_for_test(testdata / 'nonmem' / 'secondary_parameters' / 'run1.mod')
     rng = np.random.default_rng(5678)
-    stats = calculate_individual_parameter_statistics(model, 'CL/V', rng=rng)
+    stats = calculate_individual_parameter_statistics(
+        model,
+        'CL/V',
+        model.modelfit_results.parameter_estimates,
+        model.modelfit_results.covariance_matrix,
+        rng=rng,
+    )
     assert stats['mean'][0] == pytest.approx(0.0049100899539843)
     assert stats['variance'][0] == pytest.approx(7.391076132098555e-07)
     assert stats['stderr'][0] == pytest.approx(0.0009425952783595735, abs=1e-6)
 
     covmodel = load_model_for_test(testdata / 'nonmem' / 'secondary_parameters' / 'run2.mod')
     rng = np.random.default_rng(8976)
-    stats = calculate_individual_parameter_statistics(covmodel, 'K = CL/V', rng=rng)
+    stats = calculate_individual_parameter_statistics(
+        covmodel,
+        'K = CL/V',
+        covmodel.modelfit_results.parameter_estimates,
+        covmodel.modelfit_results.covariance_matrix,
+        rng=rng,
+    )
     assert stats['mean']['K', 'median'] == pytest.approx(0.004526899290470633)
     assert stats['variance']['K', 'median'] == pytest.approx(2.95125370813005e-06)
     assert stats['stderr']['K', 'median'] == pytest.approx(0.0018170955599868073, abs=1e-6)
@@ -79,7 +97,12 @@ def test_calculate_individual_parameter_statistics(load_model_for_test, testdata
 def test_calculate_pk_parameters_statistics(load_model_for_test, testdata):
     model = load_model_for_test(testdata / 'nonmem' / 'models' / 'mox1.mod')
     rng = np.random.default_rng(103)
-    df = calculate_pk_parameters_statistics(model, rng=rng)
+    df = calculate_pk_parameters_statistics(
+        model,
+        model.modelfit_results.parameter_estimates,
+        model.modelfit_results.covariance_matrix,
+        rng=rng,
+    )
     assert df['mean'].loc['t_max', 'median'] == pytest.approx(1.5999856886869577)
     assert df['variance'].loc['t_max', 'median'] == pytest.approx(0.29728565293669557)
     assert df['stderr'].loc['t_max', 'median'] == pytest.approx(0.589128711884761)
@@ -93,7 +116,12 @@ def test_calc_pk_two_comp_bolus(load_model_for_test, testdata):
     # Results are not verified
     model = load_model_for_test(testdata / 'nonmem' / 'models' / 'mox_2comp.mod')
     rng = np.random.default_rng(103)
-    df = calculate_pk_parameters_statistics(model, rng=rng)
+    df = calculate_pk_parameters_statistics(
+        model,
+        model.modelfit_results.parameter_estimates,
+        model.modelfit_results.covariance_matrix,
+        rng=rng,
+    )
     # FIXME: Why doesn't random state handle this difference in stderr?
     df.drop('stderr', inplace=True, axis=1)