+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Changes
 * sample_parameters_from_covariance_matrix will need parameter_estimates and covariance_matrix as explicit arguments
 * sample_parameters_uniformly will need parameter_esimtates as explicit arguments
 * sample_individual_estimates will need individual_estimates and individual_estimates_covariance as explicit arguments
+* calculate_individual_parameter_statistics and calculate_pk_parameters_statistics will need parameter estimates and covariance matrix
 
 Bug fixes
 =========